// src/interfaces/commonInterfaces.ts
export interface OTP {
  otp: string | undefined;
  email: string;
  otpSetTimestamp: number | undefined;
}
// ...
export interface VendorSession {
  otpSetTimestamp: number | undefined;
  email: string;
  password: string;
  name: string;
  city: string;
  contactinfo: string;
  companyName: string;
  about: string;
  otpCode: string | undefined;
  otpExpiry: number;
  resendTimer: number;
  portfolioImages?: string[];
  aadharImages?: string[];
}